+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fixed unique shiritori word bonus always being applied
+- Fixed /leaderboard not including bank balance
 
 ## [1.1.0] - 2022-09-09